+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# 0.0.1 (2023-09-22)
+# 0.1.0 (2023-09-30)
+
+### feat:
+
+- add dateT validation (#915ffab)
+- validation for primitives, objects and array (#9f7dbc5)
+
+### fix:
+
+- no need for allOf (#cb1361e)
 
 ### other:
 
 - initial commit (#f928658)
 
-### wip:
+### test:
 
-- wip: (#f7fd8cf)
+- increase coverage with objectT RangeError test (#de018d8)